refactor redirect if logged out
