+++ frontend/auth/pages/index.tsx
@@ -1,22 +1,16 @@
 import type { NextPage } from 'next'
-import Router from 'next/router'
-import { NextPageContextWithSession } from './_app'
+import {
+  getInitialPropsRedirectIfLoggedOut,
+  NextPageContextWithSession,
+} from '../src/session'
 
 const Home: NextPage = () => {
   return null
 }
 
-Home.getInitialProps = async ({ res, session }: NextPageContextWithSession) => {
-  if (!session) {
-    if (res) {
-      res.writeHead(307, { Location: '/login' })
-      res.end()
-    } else {
-      Router.replace('/login')
-    }
-  }
-
-  return Promise.resolve({})
+Home.getInitialProps = async (ctx: NextPageContextWithSession) => {
+  getInitialPropsRedirectIfLoggedOut(ctx)
+  return {}
 }
 
 export default Home
